fix: move async logic to thunk
